refactor(plugin-random): Move API types alongside Random class
